requirements.txt
loguru
selenium
python-telegram-bot
flask
apscheduler
flask-login
Flask-SQLAlchemy
passlib
wtforms
webdriver-manager
cryptography